Rewrote passphrase generation to use ssh-agent
Since it is not possible to just retrieve the SSH RSA secret key from
ssh-agent, rewrote passphrase generation to hash the public parameters
(n and e) of the RSA key, then ask the ssh-agent to sign them (which
requires the secret key). If the agent is not running, or it can't sign
(probably because the key hasn't been added to the agent), retrieves the
secret key from the id_rsa file and signs the hash.

This required a fairly big refactor of code, since the public RSA key
has two parameters (n and e), and previously, I was only passing around
the secret parameter d.