Fixed problems with passphrase generation.
Refactored to fetch public keys and private keys separately, so that
when downloading a file, it wouldn't try to access the secret key file
before it figured out that the incoming file wasn't encrypted.

In the process, discovered problem with passphrase generation. The
passphrase generated was different if ssh-agent was running than if it
was not, and we read the secret key file directly then did the RSA
signature. Turns out that the system ssh-agent ignored the requested
"rsa-sha2-256" algorithm and just did "ssh-rsa", which is a SHA1 hash.
Finally gave up and just requested "ssh-rsa" from both ssh-agent and
local signing, so they should always match.