Refactor build, ICL modifications.
Reworked changes to existing code to be inside #ifdef IRONCORE, so that
it is possible to build the original sftp and sftp-server executables
along with ironsftp (in part to facilitate testing). Corresponding
changes to the Makefile to define IRONCORE when compiling files to build
iron executables.

Also added warning if a .iron file is downloaded but can't be encrypted
(because the file is not encrypted, or it isn't encrypted to the current
user.