Change naming of decrypted files.
Instead of restoring the output of decryption to a file with the name of
the original that was encrypted, just strip the .iron extension from (if
there is one), or add a .decrypt extension to the name of the downloaded
file (if it was encrypted). It does still output an info message if the
output file name is different from the name of the file that was
originally encrypted.