Display lock icon on files being uploaded.
If you do "put abc", ironsftp was displaying abc in the progress meter,
and because the file name did not end in .iron, it didn't display the
lock icon. Fixed so that it displays "abc.iron" as the name of the file
being uploaded, which displays the lock icon.

Also restat'd the file after it was encrypted, so the file size properly
matched the file that was actually being uploaded.