Modified handling of file names.
Now, write_gpg_encrypted_file generates an output file name by
appending .iron to the input file. If that file already exists, it
generates a file with a .iron extension using mkstemps.
write_gpg_decrypted_file strips the .iron from the input file and opens
that name as the output file. It will overwrite the file if it exists.
An ironsftp put will always append .iron to the remote file name (if it
doesn't already have it). If the remote file exists, prompts to confirm
overwrite.
An ironsftp put will always append .iron to the local file name to form
the download file name. If that name exists, prompts to confirm
overwrite. It also prompts to confirm overwrite of the file without the
.iron extension. After the .iron file is downloaded, if it was not
encrypted, renames it from the .iron file to the output file name.