Don't prompt for passphrase if not needed
If the SSH secret key file doesn't have a passphrase, don't prompt for
one. (Attempt to load the file with no passphrase first, and then prompt
if that fails).
Also cleaned up code to copy the SSH private and public key files - was
doing it every time it needed to access the private key file, instead of
on key generation. Also, read from the copied key files instead of the
main files, in case someone generates a new id_rsa file after we
generate iron keys.