Fix public key handling.
When generating pubkey file, create as ~/.ssh/ironpubkey. Then upload
to server as ~/.ironpubkey. When adding a recipient, instead of getting
the pubkey from their home directory on local machine, retrieve it from
the remote machine and store it in ~/.ssh/pubkeys/ironpubkey.<login>.

This sets up to create the directory of public keys in the future.

Since we no longer need to figure out other users' .ssh directory paths,
refactored code to eliminate a lot of ssh_dir parameters, and when we
need it for the current user, to use iron_user_ssh_dir.