Create key ID index to find user names / pubkeys
Any time we fetch a pubkey, generate an index entry that maps the
key ID to the user@host name of the pubkey.
Added commands finduser (to fetch pubkey without adding recipient),
clrrcpt (to clear recipients), showrcpt (to show current list of
recipients).
When downloading an encrypted file, lists the logins of the users where
key IDs can be resolved using the index. Lists the count of key IDs that
could not be resolved.