Fix signature calculation problem. (#43)
If an RSA signature is computed for a hash and at least eight of the
high order bits are zero, when the signature gets converted to a bignum,
those bits get dropped. When that happens, the length of the data in the
data signature packet ends up being shorter than the length we put in
the header. That was detected and reported as an "internal error".

To fix, removed that check, and instead added padding of zero bytes to
make the length of the data signature body match the size in the packet
header. It was necessary to do this rather than just write the extra
byte(s) of zeros at the start of the Multi-Precision Integer (MPI)
because then libgcrypt hated it. However, after doing the padding, the
signature woudln't verify, because the OpenSSL RSA verification function
won't accept any signature whose length doesn't match the RSA key
length. So on verify, I had to make a temporary buffer and pad the start
with zeroes to make it match the RSA key length.

Note that this wasn't a problem with the Public Key / UID/ Signature
packet signature, because we didn't precalculate the length of the
packet to write the header before streaming, and because we don't ever
verify that signature currently in the irongpg code.