Increase max key size
If the high bit of an RSA 4096-bit key is 1, the key actually takes up
513 bytes (for the leading 0 padding for the sign bit). Increased the
max key length to handle RSA 8192-bit keys with an extra padding byte.
Fixed a couple other places in the code where a buffer was not long
enough.
Regenerated the test keys to be RSA 4096-bit instead of 2048-bit.