Switch to ed25519 signing key - issues 3, 55 (#56)
* Issues 3, 55. New Ed25519 signing keys

Instead of requiring an id_rsa SSH key, then using that to sign files
and keys, allow the use of any SSH key, and generate a new Ed25519 key
when initializing ironsftp. The only time RSA keys are used other than
to generate the passphrase to access ed25519/curve25519 secret keys is
to validate the signature of old files uploaded with the first version
of ironsftp.